some changes in code.

+++ app/managers/inventory_manager.py
@@ -1,13 +1,15 @@
-from app.models.product import CreateProduct, Products
-from app.models.user import Users
+from app.models.product import CreateProduct
 from sqlmodel import Session, select
-from fastapi import HTTPException, status
-from app.utils.jwt import verify_token
+from fastapi import HTTPException, status                                                                                                                                                                                                                                   
+from app.models.user import Users
+from app.models.product import Products
 
 class InventoryManager:
-    def add_products(self, product: CreateProduct, session: Session)->str:
+    def add_products(self, product: CreateProduct, user_token: dict, session: Session):
+      db_user = session.exec(select(Users).where(Users.id == user_token["id"])).first()
+      
       new_product = Products(
-        user_id = product.user_id,
+        user_id = db_user.id,
         name = product.name,
         brand = product.brand,
         price = product.price,
@@ -22,10 +24,6 @@ class InventoryManager:
       return "Product successfully save."
 
     def show_all_products(self, user_token: dict, session: Session):
-        db_users = session.exec(select(Users).where(Users.id == user_token["id"])).first()
-        if not db_users:
-           raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized user!")
-        
         db_products = session.exec(select(Products).where(Products.user_id == user_token["id"])).all()
 
         if not db_products:
@@ -34,25 +32,31 @@ class InventoryManager:
         return {"Products": db_products}
       
 
-    def search_product(self, product_id: int, session: Session)->str:
+    def search_product(self, product_id: int, user_token: dict, session: Session):
         db_product = session.exec(select(Products).where(Products.id == product_id)).first()
 
         if not db_product:
-           raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found!")
+          raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found!")
         
+        if db_product.user_id != user_token["id"]:
+           raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized user!")
+           
         return db_product
 
-    def update_stock_data(self, product_id: int, stock_quantity: int, session: Session)->str:
+    def update_stock_data(self, product_id: int, stock_quantity: int, user_token: dict, session: Session):
       if product_id <= 0:
         return "Invalid product_id!"
-      
+
+      if stock_quantity <= 0:
+        return "Invalid stock quantity!"
+        
       db_product = session.exec(select(Products).where(Products.id == product_id)).first()
       
       if not db_product:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found!")
-
-      if stock_quantity <= 0:
-        return "Invalid stock quantity!"
+      
+      if db_product.user_id != user_token["id"]:
+         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized user!")
       
       db_product.available_stock += stock_quantity
 
@@ -62,7 +66,7 @@ class InventoryManager:
       return{"message": f"Successfully add {stock_quantity} pieces in stock."}
       
 
-    def delete_product(self, product_id: int, session: Session)->str:
+    def delete_product(self, product_id: int, user_token: dict, session: Session)->str:
         if product_id <= 0:
           return "Invalid product_id"
         
@@ -70,6 +74,9 @@ class InventoryManager:
 
         if not db_product:
            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found!")
+        
+        if db_product.user_id != user_token["id"]:
+           raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized user!")
 
         session.delete(db_product)
         session.commit()
